Moved store logic from Landing to Dashboard to account for page refresh
